backend/requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
google-generativeai
python-dotenv
supabase
python-multipart
python-docx
